dmapagent/tests/test_workflow.py
"""Integration tests for workflow."""
import pytest
from src.graph.workflow import run_mapping_workflow
from src.mapping.mapping_lineage import MappingType


@pytest.mark.integration
def test_workflow_with_sample_data():
    """Test complete workflow with sample data."""
    # This test would require sample DDL and Excel files
    pass


@pytest.mark.integration  
def test_multi_source_mapping():
    """Test mapping from multiple sources."""
    pass


@pytest.mark.integration
def test_multi_destination_mapping():
    """Test mapping to multiple destinations."""
    pass
