dmapagent/requirements.txt
langchain
langgraph>=1.0.0
google-generativeai
anthropic
ollama
openpyxl==3.11.0
pandas>=2.2.0
pydantic>=2.6.0
pydantic-settings
python-docx==1.0.0
beautifulsoup4==4.12.0
pdfplumber==0.10.0
sqlparse==0.5.0
python-dotenv==1.0.0
typing-extensions==4.9.0
httpx==0.25.0
